+++ src/player/setting.h
@@ -34,8 +34,8 @@ public:
     bool offensive_kick_planner_use_sample_pass = true;
     bool offensive_kick_planner_use_sample_dribble = true;
 
-    int moving_save_energy = 60; // 0 to 100
-    int moving_pressing_level = 50; // 0 to 100
+    double moving_save_energy = 60; // 0 to 100
+    double moving_pressing_level = 50; // 0 to 100
     bool moving_use_offside_trap = true;
 
     void find_version(const string &json_string);
